requirements.txt
asgiref==3.8.1
certifi==2025.1.31
channels==4.2.0
charset-normalizer==3.4.1
Django==5.1.7
djangorestframework==3.15.2
idna==3.10
requests==2.32.3
sqlparse==0.5.3
urllib3==2.3.0
